:bug: unexpected signed expansion

+++ src/bdm/device/spi/spidev.cxx
@@ -26,8 +26,8 @@ namespace bdm::device::spi {
 
 	void spidev::transfer(std::uint32_t size, const std::uint8_t* tx, std::uint8_t* rx) {
 		spi_ioc_transfer tr = {};
-		tr.tx_buf = reinterpret_cast<std::uint64_t>(tx);
-		tr.rx_buf = reinterpret_cast<std::uint64_t>(rx);
+		tr.tx_buf = reinterpret_cast<std::size_t>(tx);
+		tr.rx_buf = reinterpret_cast<std::size_t>(rx);
 		tr.len = size;
 		tr.speed_hz = _conf.max_speed;
 		tr.bits_per_word = _conf.bits_per_word;
